requirements.txt
cryptography>=41.0.0
Pillow>=10.0.0
plyer>=2.1.0
openpyxl>=3.1.0
watchdog>=3.0.0